📦 updating to Killua 0.12 (#569)
* Added Patreon checks

What has been tested:
+ the Patreon class returning what it's supposed to, tested thanks to @Flicksie to work with even hundreds of Patrons

What has not been tested:
+ The badges checks working
+ The booster role check working
+ Badges being successfully added and removed

* Deleting unnecessary code

* Fixing Patreon update, improving paginator

* Fixed boosts not giving premium, nitro badge on `k!profile` being displayed wrong

* Fixing k!novel + paginator

Fixed an issue with the novel command that would make it not paginate-able, then fixed the issue with the paginator ignoring results from `func`. I also significantly improved the k!novel code in readability

* replacing `avatar.replace(format="png")` with str(avatar.url)`

Plus accidentally adding api.py

* Adding game, improving other code

Revisit the code for `rps` and adding a pretty smart formula to it. Also changing the design to use interactions

Adding a `count` game with difficulty `easy` and `hard`

Adding a new ipc route

Adding a ConfirmButton class that will be used in the future instead of waiting for the user to confirm or deny by sending "y" or "n".

every action has been modified to never exceed maximum embed title length, so you can hug unlimited people now

The user profile now uses fields instead of the embed description

Interaction ion which `on_interaction` returns are now defered since my pr adding that behaviour was denied

* Adding pypxl 0.2.1 functionality

* Doing some major premium improvements + other stuff

+ Premium badges are now saved as the id Patreon gives them to make things easier

+ Patreons can now choose x servers (depending on tier) to make a premium server. Transferring that has a cooldown of 7 days

+ Edited documentation in todo system and replaced a [y/n] confirm message with the new ConfirmButton class

+ fixed an issue where the ConfirmButton class wouldn't disable its buttons on a timeout

+ Added more constants to the constants file coming from config.json

+ Put all premium related stuff in a separate file

+ Update the User class accordingly to the new way premium subscriptions are saved

* One of the biggest commits in this version

This commit adds, fixes and changes a lot of things

__Added stuff__
+ added lootboxes. LootBoxes can be obtained by the new command `k!weekly` (premium only) or every 10 votes or by buying them in the shop
+ added cache for book background to make `k!book` even faster
+ added command `boxinfo` which gives info about a box
+ added `k!open` to open any box
+ added `k!inventory` displaying your boxes
+ added banner to `k!profile`. This will most likely be removed in the final version due to being required to make an API call for it

__Changes__
+ Commands are now formatted by a method of m subclass of Bot. This also fixes an issue with the help command not displaying all groups along with not being required to type out the group prefix for commands in command groups. Eg `todo create` is now only `create` and `todo` is automatically added.
+ Added that `k!8ball` will now also be displayed as `k8ball` in the help menu
+ Added a new file: `shop.py`. This file contains command groups `shop`, `buy` and `give`. These have subcommands for different shops for different parts of the bot. Doing this was a design choice to make this more understandable to readers

This update includes numerous other small fixes, changes and additions I am not listing here

* New game, minor improvements

This commit adds
+ a new game, `k!trivia`
+ a paginator to `k!usage`
+ more usage of the `ConfirmButton` class
+ better rps handling
and some more minor fixes and improvements

* :bug: Fixing paginator

Fixed a bug where the footer would stay the same across multiple paginators

* Rewrote k!hunt, fixed some todo stuff

+ rewrote k!hunt to be faster and cleaner
+ fixed some issues with the todo system like when creating or viewing a list

* finished card system, added class cache and setup.py for contributors

This commit has some of the main points of this version. Let's first look at the greed island stuff
- the use command now works like this; every spell card has a subclass of Card with an exec method that contains what to do in
  case that card is used. then a function looks for that class and method when you use the command and checks the annotations on
  the use command with what is provided to see if it's valid and if it is executes it. I found this to be the best solution
- added reward when reaching a full book! You now get the card 0 and a special badge as long as your book is full. These will be r
   removed if it stops being full tho!
- all cards are loaded into the cache on startup now with it printing the progress

Then I made some improvements at all the classes in general
- There is a cache added to the User, Guild and TodoList cache to minimise db requests
- Improved a lot of design in the User class

Fixes
- fixed an issue where it would sometimes remove premium guilds from boosters
- fixed some issues with the todo lists

There is also a new file called setup.py which can be run to set up the right db layout for new contributors. the README reflects that now, furthermore is stuff now printed in color

* Bug fixes, improvements

A lot of todo bugs were fixed, the paginator now supports subclassing and the help command allows you to return to it

* Small things so I don't forget them

* Fixing lgtm alerts

Also small improvements

* Fixing remaining lgtm alerts and `update` and `publish_update`

* Adding premium stuff and doubled jenny as a patreon tier reward

* First try at an issue template

* Bug template

* Feature request template

* Create config.yaml

* Create pull_request_template.md

* Fixing minor issues, adding error message

* Bug fixes thanks to my testers :)

* Fixed typos (#570)

* Fixed typo

* Actual spelling fixes

This doesn't only fix mistakes in strings/comments but also 2 cases where a variable was misspelled (tuser & scontext)

* Rewriting more to cache

Made the cooldown and dashboard check use the cache
Improved the design of the cooldown message

* Bug fixes, adding more boxes, small improvements

* Small bug fix, fix lgtm alert, edit constants

* Fixing bugs, adding lootboxes and command

- Fixed that you could steal or destroy or do whatever with card 0
- added k!spin which spins the provided image in a circle
- added `gif.py` a whole file dedicated to that command because pillow is bad
- renamed `pxlapi.py` -> `image_manipulation.py`
- fixed numerous other bugs

* Adding more icons, fixing k!spin

* Only using intents I need

* Adding helper send method more often, fixing lgtm alert, adding badge

* Fixed a small but dangerous bug

* Fixed bugs, added some texts

* improving paginator, fixing bug, adding last missing badge, added texts

* Woops forgot to reverse something for testing

* Fixing buges, adding features

Fixed some bugs
Disabled skip to end/beginning when you would end up on the same page
Added k!check <id> to check how many copies of a card you own are fake

* :bug:

* Final changes

Polished the mute command
Added more texts
Improved logic of k!discard
Improved k!server
Fixed playing rps against Killua

* Small improvements

Took a second look at the ban command
Added helper `send_message` function the first time a paginator send a message
Added 100+ more topics

* Improving Imagefunction (#572)

Making cards transparent and adjusting their position

* Fixing lgtm alert, adding more text

* Fixing some book positions and book bug

* Final commit

Improving book command once again
adding shop artwork
improving `remove_card`

+++ killua/checks.py
@@ -1,9 +1,24 @@
+
 import discord
 from discord.ext import commands
-from .constants import blacklist
+from typing import Union, Type
 
-cooldowndic = {}
+from .constants import blacklist, PatreonBanner
+from .classes import User, Guild
 
+cooldowndict = {}
+
+def _clean_command_name(command:Union[commands.Command, Type[commands.Command]]) -> str:
+	"""returns the clean command name of a command"""
+	if not command.parent:
+		return command.name
+	else:
+		name = ""
+		c = command
+		while c.parent:
+			name = c.parent.qualified_name + " " + name
+			c = c.parent
+		return name + command.name
 
 def blcheck(userid:int): # It is necessary to define it twice as I might have to use this function on its own
     """
@@ -31,12 +46,13 @@ def check(time:int=0):
     from datetime import datetime, timedelta
     from killua.constants import guilds, teams, blacklist, stats
 
-    def add_usage(command:str):
+    def add_usage(command:Union[commands.Command, Type[commands.Command]]) -> None:
         data = stats.find_one({'_id': 'commands'})['command_usage']
+        command = _clean_command_name(command)
         data[command] = data[command]+1 if command in data else 1
         stats.update_one({'_id': 'commands'}, {'$set': {'command_usage': data}})
 
-    def blcheck(userid:int):
+    def blcheck(userid:int) -> bool:
         """
         Input:
             userid (int): The id of the user who should be checked
@@ -45,7 +61,7 @@ def check(time:int=0):
             (boolean): if the user is blacklisted or not
 
         Purpose:
-            Checking before everry command if the user is blacklisted
+            Checking before every command if the user is blacklisted
         """
 
         result = blacklist.find_one({'id': userid})
@@ -55,75 +71,57 @@ def check(time:int=0):
         else:
             return True
     
-    async def custom_cooldown(ctx, time:int):
-        global cooldowndic
+    async def custom_cooldown(ctx, time:int) -> bool:
+        global cooldowndict
         now = datetime.today()
-        later = datetime.now()+timedelta(seconds=time)
         try:
-            cdwn = cooldowndic[ctx.author.id][ctx.command.name]
-        except KeyError as e:
+            cdwn = cooldowndict[ctx.author.id][ctx.command.name]
+        except KeyError as e: # if there is no entry in the cooldowndict yet for either the command or user
             error = e.args[0]
             if error == ctx.author.id:
-                cooldowndic = {ctx.author.id: {ctx.command.name: later}}
+                cooldowndict = {ctx.author.id: {ctx.command.name: now}}
                 return True
             if error == ctx.command.name:
-                cooldowndic[ctx.author.id][ctx.command.name] = later
+                cooldowndict[ctx.author.id][ctx.command.name] = now
                 return True
 
-        cd = cdwn-now 
+        cd = now-cdwn 
+        diff = cd.seconds
+        
+        user = User(ctx.author.id)
+        guild = Guild(ctx.guild.id) if ctx.guild else None
+        view = discord.ui.View()
+        view.add_item(discord.ui.Button(label="Get premium", url="https://patreon.com/kilealkuri", style=discord.ButtonStyle.blurple)) # sadly I cannot color a link button :c
 
-        if str(cdwn) < str(now):
-            cooldowndic[ctx.author.id][ctx.command.name] = later
+        if guild and guild.is_premium:
+            time /= 2
+
+        if user.is_premium:
+            time /= 2
+
+        if diff > time:
+            cooldowndict[ctx.author.id][ctx.command.name] = now
             return True 
-
-        else:
-            user = teams.find_one({'id': ctx.author.id})
-            guild = guilds.find_one({'id': ctx.guild.id}) if ctx.guild else None
-
-            if cd.seconds < time:
-                t = -1*(6-time-cd.seconds)
-                if guild is None:
-                    pass
-                elif 'partner' in guild['badges'] or 'premium' in guild['badges']:
-                    if int(cd.seconds) > time/2:
-                        t = t/2
-                    else:
-                        cooldowndic[ctx.author.id][ctx.command.name] = later
-                        return True
-
-                if user is None:
-                    await ctx.send(f':x: Command on cooldown! Try again after `{t}` seconds\n\nHalf your cooldown by becoming a patreon here: https://patreon.com/kilealkuri', delete_after=5)
-                    return False
-
-                if 'premium' in user['badges']:
-                    if int(cd.seconds) > t/2:
-                        await ctx.send(f':x: Command on cooldown! Try again after `{t/2}` seconds\n\nHalf your cooldown by becoming a patreon here: https://patreon.com/kilealkuri', delete_after=5)
-                        return False
-                    else:
-                        cooldowndic[ctx.author.id][ctx.command.name] = later
-                        return True
                     
-                await ctx.send(f':x: Command on cooldown! Try again after `{t}` seconds\n\nHalf your cooldown by becoming a patreon here: https://patreon.com/kilealkuri', delete_after=5)
-                return False
+        await ctx.send(f':x: Command on cooldown! Try again after `{time-diff}` seconds\n\nHalf your cooldown by clicking on the button and becoming a Patreon',file=PatreonBanner.file(), view=view, delete_after=10)
+        return False
+
+    async def settings_check(ctx) -> bool:
+        if not ctx.guild:
             return True
 
-    async def settings_check(ctx):
-
-        guild = guilds.find_one({'id': ctx.guild.id})
-
-        if not 'commands' in guild:
-            return True
+        guild = Guild(ctx.guild.id)
             
-        if not ctx.command.name in guild['commands']:
+        if not ctx.command.name in guild.commands:
             return True
 
-        command = guild['commands'][ctx.command.name]
+        command = guild.commands[ctx.command.name]
 
         # Checking if a command is disabled
         if command['enabled'] is False:
             return False
 
-        # Checking if the member is whitelisted (not implemeted)
+        # Checking if the member is whitelisted (not implemented)
         #if ctx.author.id in [int(x) for x in command['restricted_to_members'] if len(command['restricted_to_members']) > 0]:
             #return True
 
@@ -136,7 +134,7 @@ def check(time:int=0):
             return False
 
         # Checking if the channel is whitelisted if it is only whitelisted to a few channels
-        if not ctx.channel.id in command['restricted_to_channels'] and len(command['restricted_to_channels']) > 0:
+        if not ctx.channel.id in command['restricted_to_channels']:
             return False
 
         # Checking if the user has a role the command is restricted to
@@ -148,8 +146,11 @@ def check(time:int=0):
             return False
 
         # Checking if the command invokation should be deleted
-        if command['delete_invokation']:
-            await ctx.message.delete() # This could break stuff, fixing later
+        if "delete_invokation" in command and command['delete_invokation']:
+            try:
+                await ctx.message.delete()
+            except discord.HTTPException: # if it is already deleted for some reason
+                pass
 
         return True
 
@@ -160,62 +161,18 @@ def check(time:int=0):
         try:
             if (await settings_check(ctx)) is False:
                 return False
-        except Exception: #If someone used the api and messed up the guilds data struture
+        except Exception: #If someone used the api and messed up the guilds data structure
             pass
 
         if time == 0:
-            add_usage(ctx.command.name)
+            add_usage(ctx.command)
             return True
 
         if await custom_cooldown(ctx, time) is False:
             return False
 
-        add_usage(ctx.command.name)
+        add_usage(ctx.command)
 
         return True
 
     return commands.check(predicate)
-
-'''async function p
-Input:
-self: taking in self because it is outside of a function
-
-Returns:
-Nothing
-
-Purpose:
-Changing Killuas presence freqently if he is added to a guild, removed or 12 hour pass, now also customisable by me at will
-'''   
-
-async def p(self):
-    from killua.constants import presence
-    from datetime import datetime, timedelta, date
-
-    status = presence.find_one({})
-    if status['text']:
-        if not status['activity']:
-            status['activity'] = 'playing'
-        if status['activity'] == 'playing':
-            s = discord.Activity(name=status['text'], type=discord.ActivityType.playing)
-        if status['activity'] == 'watching':
-            s = discord.Activity(name=status['text'], type=discord.ActivityType.watching)
-        if status['activity'] == 'listening':
-            s = discord.Activity(name=status['text'], type=discord.ActivityType.listening)
-        if status['activity'] == 'competing':
-            s = discord.Activity(name=status['text'], type=discord.ActivityType.competing)
-
-            
-        if not status['presence']:
-            status['presence'] = 'online'
-        if status['presence'] == 'online':
-            return await self.client.change_presence(status=discord.Status.online, activity=s)
-        if status['presence'] == 'dnd':
-            return await self.client.change_presence(status=discord.Status.dnd, activity=s)
-        if status['presence'] == 'idle':
-            return await self.client.change_presence(status=discord.Status.idle, activity=s)
-    a = date.today()
-    #The day Killua was born!!
-    b = date(2020,9,17)
-    delta = a - b
-    playing = discord.Activity(name=f'over {len(self.client.guilds)} guilds | k! | day {delta.days}', type=discord.ActivityType.watching)
-    return await self.client.change_presence(status=discord.Status.online, activity=playing)
